backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
openpyxl==3.1.5
google-generativeai==0.7.2
openai==1.40.0
bson==0.5.10
